+++ tests/testthat/setup.R
@@ -4,4 +4,4 @@ con <- CoreAPIV2::authBasic(api, useVerbose = verbose)
 data <- jsonlite::fromJSON(env$data)
 
 # configure test environment for JSON SDK usage
-CoreAPIV2::loadXmlConfigFile(con$coreApi, appName = "core_CoreApp.SDK")
+CoreAPIV2:::loadXmlConfigFile(con$coreApi, appName = "core_CoreApp.SDK")